refactor: jpa로 마이그레이션 준비
