+++ src/main/java/kitchenpos/domain/vo/Price.java
@@ -2,10 +2,17 @@ package kitchenpos.domain.vo;
 
 import java.math.BigDecimal;
 import java.util.Objects;
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
 
+@Embeddable
 public class Price {
 
-    private final BigDecimal value;
+    protected Price() {
+    }
+
+    @Column(name = "price")
+    private BigDecimal value;
 
     private Price(final BigDecimal value) {
         this.value = value;
