+++ src/main/resources/application.properties
@@ -2,3 +2,4 @@ logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
 spring.h2.console.enabled=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=validate
